Split page history to component
