+++ framerail/src/routes/[slug]/[...extra]/index.ts
@@ -1,5 +1,6 @@
 export { default as EditorPane } from "./EditorPane.svelte"
 export { default as FilePane } from "./FilePane.svelte"
+export { default as HistoryPane } from "./HistoryPane.svelte"
 export { default as LayoutPane } from "./LayoutPane.svelte"
 export { default as MovePane } from "./MovePane.svelte"
 export { default as ParentPane } from "./ParentPane.svelte"
